Fixed argc, argv, & envp thanks to rian@HN's suggestions.  Also removed 01__init (one file startup code!).  Makefile, README, and TODO updates.  101 sloc. :/

+++ lib/_exit.c
@@ -7,7 +7,7 @@
 
 void _exit( int r )
 {
-   int ret = syscall1( SYS_exit, r );
+   long ret = syscall1( SYS_exit, r );
    ret = syscall3( SYS_write,
                    1,
                    ( long )( "SYS_exit failed.\n" ),
